adding CSS feats for visual aid
TODO

FE: Need to import and work on fetching api requests.

BLOCK on BE - Issue with creating duplicate assets within 1 portfolio

BLOCK on BE- Datetime needs to be uniform from front to end - I've added a helper function while debugging to see what's happening but it's no where near a fix.

TODO FE: Need to implement edit button to change asset info

TODO FE: When creating an asset, if asset not in list, we will receive a 404 error on FE. When we receive this, we need to have a logic to enable the user to add their a manual asset.